Task 1 - Accepting more inputs
I was able to make some changes to the program by allowing more input
options, rather than just Y and N. It will now accept "y", "n", "Y",
"N", "Yes" and "No". And the way I did this was by using the capitalize
function in python that capitalizes the first letter of the string. So
now a "y" will turn into a "Y", and this made it easier so there doesn't
have to be as big a line of coding for the while loop