Task 2 - Accepting more inputs for the main menu
This task was similar to the task before but instead of changing the
inputs of the game itself, it was the main menu. This was much the same,
as you add the .capitalize() function at the end of the return function
and widen the while loop so it accepts "Q" and "Quit" instead of just
"q".