Task 3b - Asking player for leaderboard
In this question, you had to ensure that the user had the option to
either enter their name onto the leaderboard or not. When I did this, I
made it a simple if statement, where if the user enters for yes, then
their score goes on, but anything else and their score doesn't go on. I
did this coding in the PlayGame function so that it could potentially
skip over a lot of unnecessary coding inside the other functions.
As for the other changes, I realised that that the game itself wasn't
working because the code was always assuming that when the user made a
decision, it would always be wrong. So to tackle this, I've ensured that
the coding for when the user guesses above is True, and below for False,
making it easier to navigate and understand for the program