Task 4 - Formatting the Output
For this task, I had to ensure that the DisplayRecentScores() function
was outputting the scores in a more relevant form. This means that the
name and table are formatted into an easy-to-read format. The way I did
this, was I used the {0:10} format. The 0 is the value that is going to
be added to the front, in this case the name of the player, and the 10
means that the name cannot pass that point. The second point one,
{1:<15}, ensures that the second variable (the score) starts at 15 paces
in front.