Task 5 - Introducing date and time to leaderboard
So for task 5, we were asked to add to the leaderboard the date and time
that someone finished the scores. So what happened was that to do this,
I introduced a new function called GetCurrentDate() which would go
through and get the date. This worked by using the
datetime.datetime.now(), which would be useful in getting the current
minute/hours/etc. This is because right after you name the
datetime.datetime.now(), you can now get the certain variables in that
easily, and so I got those, then added them all together into
Current_Time, and passed on the variable.
The other changes that I made to the program was that I added a
self.Date, which would be easy enough to edit at any given time, and
also ensured that it was easy to go through when passing the variable
Date.