Task 6. Part 3
So the task finally got to work instead of just the options menu. The
way I was able to change the Ace as 1 into Ace as 14 was by changing the
function: IsNextCardHigher. What this means is that I don't change the
original value of the deck but instead if the rank number starts as a 1,
inside the function it will change the value straight to 15, above the
value of the king.
One of the problems that I had was being able to pass the value of High
through the program, and that's how the pdb helped. It showed me that
the value High wasn't going through the value but instead was being
stopped. This is because I forgot to create a return value that would've
meant that the value goes through, instead of receiving an empty
variable.