Task 7. Part 2
Struggling to cope with the bubble sort coding. I have taken the coding
of a bubble sort and applied it to the deck of cards, but apparently it
doesn't work, as it will either not show a score and date or it will
just do the same thing with the scoring system. I managed to find 1
large change, which is that when comparing, you compare the score and
not the entire variable