Task 7. Part 3 - Completed
This is the finalised version of the bubble sort mechanic in the deck of
cards. There were a couple of problems that I was able to change and add
to make the program work. Firstly, I needed to change the moreSwaps
variable into True as it wasn't going through the code.
Secondly, I removed the + 1 o the range(1, NO_OF_RECENT_SCORES +1)
Finally, since RecentScores is passed through reference, I didn't need
to do a return