Task 9 - Formatting problem
For task 9, I have managed to bring back the data from the other games
and put them into the program through the LoadGame function. But, I
currently have a formatting problem, which results in the DisplayScore()
showing the data as spaced out, not all on the same line and very messy.
This is probably due to the fact that when I transferred the data over
the file, it was taken with a "\n" at the end, which is used to make a
new line in the text file. Somehow, I need to either remove this extra
bit or make it so that it doesn't affect the actual formatting of the
code.