Task 9 Complete, but still with errors
I have completed the task 9, by being able to retrieve information from
the text file. But I haven't been able to make sure that it detects
errors in the coding. For example, when you load the files, it seems to
be safe, but as soon as you input your name after a game, there is an
error which claims to be unordable, and problems with converting str to
int, even though I thought that there wasn't any strings in the coding