Graphical interface to field simulation program
In this one, I added the graphical interface to the field simulation
program, and as a result was able to add some widgets to my coding, such
as radio buttons, text boxes, submit buttons and images.
Now I just need to debug the coding to ensure that the coding is working
correctly.