Task 2a - Adding the radio button widget class
Adding the radio button widget class, this will allow for the radio
buttons to be added to the program in due time. I also changed the name
of the file Crop Simulator (redo).py to Crop_Simulator_(redo).py, as
when importing data from each set of coding, you can't have gaps
inbetween the words.