Task 5c - Crop Simulator Complete
I've added to the code. I've added the ability to add images, and I've
managed to get a fully functioning user interface and have managed to
change the images that show on the screen. All I need to do now is get
those images onto the same computer (they're at home) and just commit
them to the folder so that it fully works now.